requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
pydantic==1.10.11
pyyaml==6.0
